backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
httpx==0.27.0
python-dotenv==1.0.1
groq==0.11.0
deepgram-sdk==3.7.0
googletrans==4.0.0rc1
gTTS==2.5.3
supabase==2.9.1
redis==5.1.0
python-multipart==0.0.12
